Criação dos models e relacionamentos

+++ src/main/java/com/br/finnet/integracaoAdiq/domain/repositories/PaymentRepository.java
@@ -4,6 +4,8 @@ import com.br.finnet.integracaoAdiq.domain.models.PaymentModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import javax.persistence.criteria.CriteriaBuilder;
+
 @Repository
-public interface PaymentRepository extends JpaRepository<PaymentModel, Long> {
+public interface PaymentRepository extends JpaRepository<PaymentModel, Integer> {
 }
